+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Writing user preferences in the purchase-info data
+
 ## [2.151.0] - 2022-04-04
 
 ### Fixed